* CMake: re-enable GCC -Wsign-conversion for GCC-10
Disabled in: ba7b9055c331f3a71732691517da8d3ac1362180
GCC 10 changes: "G++ no longer emits bogus -Wsign-conversion warnings with explicit casts."

+++ Sudoku/Sudoku/Options.h
@@ -31,7 +31,7 @@ template<int E>
 class Options
 {
 	static_assert(E >= 1);
-	using bitset = std::bitset<E + 1>;
+	using bitset = std::bitset<gsl::narrow_cast<size_t>(E + 1)>;
 
 public:
 	Options() noexcept;
